// libs/features/customer-survey/src/lib/shared/enums/cim-routes-enum.ts
/**
 * Copyright GOSI. All Rights Reserved.
 * This software is the proprietary information of GOSI.
 * Use is subject to license terms.
 */

export enum CimRoutesEnum {
  CONTRIBUTOR_SEARCH = 'home/profile/contributor/search',
  USER_SEARCH = 'home/profile/user/search',
  USER_PROFILE_INFO = 'home/profile/user/info',
  USER_UPDATE_ADDRESS = 'home/profile/user/update-address',
  CONTRIBUTOR_PROFILE_ADD_IQAMA = 'home/profile/contributor/add-iqama',
  CONTRIBUTOR_PROFILE_ADD_NIN = 'home/profile/contributor/add-nin',
  CONTRIBUTOR_PROFILE_ADD_PASSPORT = 'home/profile/contributor/add-passport',
  CONTRIBUTOR_PROFILE_ADD_BORDER = 'home/profile/contributor/add-border',
  CONTRIBUTOR_PROFILE = 'home/profile/contributor',
  CONTRIBUTOR_PROFILE_ENGAGEMENT = 'home/profile/contributor/engagement',
  CONTRIBUTOR_PROFILE_DOCUMENTS = 'home/profile/contributor/documents',
  CONTRIBUTOR_PROFILE_INJURY = 'home/profile/contributor/injury',
  CONTRIBUTOR_PROFILE_HISTORY = 'home/profile/contributor/history',
  VALIDATOR_ADD_IQAMA = 'home/profile/validator/add-iqama',
  VALIDATOR_ADD_BORDER = 'home/profile/validator/add-border',
  VALIDATOR_ADD_PASSPORT = 'home/profile/validator/add-passport'

}